Set up Wikit (vue3compat build)
Use the migration build, mapping not only vue but also
@vue/composition-api (used by Wikit) to @vue/compat. For @click.native
to work, we need MODE: 2 in the Vite config, or at least I haven’t yet
found a way to make it work in MODE: 3. At some point, we’re hoping to
have a properly Vue 3-compatible (subset of) Wikit, so that we can drop
the migration build and use Vue 3 alone.

Bug: T301145