Set up SCSS styles
Vite supports them automatically, we just need to add the right package
to our dev dependencies. My npm also decided to sort the dependencies at
the same time, which I think we might as well keep.

Also, remove the label styles, there’s no label in the HelloWorld
component (I guess it must’ve been in the original template but was
removed before the initial commit of the Vite setup).

Bug: T300891