Add i18n plugin
In the compositionApi we cannot access `this`, so using a global
instance method does not work. The established way to deal with this in
Vue 3 is to use provide/inject.

Using provide/inject works well, with the one exception that requires a
fallback or will be typed Messages|undefined. This leaves a couple of
non-ideal solutions:
1. assert it as non-null it with `!`
2. cast it with `as Messages`
3. have some if-guard that throws an Error
4. Provide a simple fallback

I went with option 4 and extended the original Messages class from Data
Bridge[0] with a fallback if no repository is provided. This also makes
the initialization in the index.html easier and still allows for a more
elaborate mock in the future, if we want it.

[0]: https://github.com/wikimedia/Wikibase/blob/f73e98507da6e494ffaa1fe9c15b787a527ba33a/client/data-bridge/src/presentation/plugins/MessagesPlugin/Messages.ts

Bug: T301570