fix Netlify hosting
Netlify actually needs an index.html in the dist directory. However this
is not added for a library build. So this creates an extra build command
to keep building it as an app that makes still use of the index.html,
link `npm run dev`. This skips the type checking as the build currently
runs on Netlify itself and there we only have a limited amount of build
minutes included. Rather the type checking should be part of the "code
quality" CI job.

This Netlify hosting was broken in #19 / 171d33f