Run type checks in CI (#34)
This ensures that types are checked in our normal CI runs. Previously,
types were only checked as part of the build script, which is not
normally run in CI.

It is moved to its own script name, because it is not a test per se and
it is way too slow for linting.