Add spacing between the components to the form
The button needs a wrapping <div> so it doesn't get stretched to the
full width. (This button really should also be its own component).

Unfortunately, we cannot use grid here, because that is not supported in
all 10.x versions of Safari and we claim to support them all.
We can also not use flex with gap because that that as well only
supports gap in version 14.1+: https://caniuse.com/flexbox-gap

Instead we're using the owl selector based on
https://every-layout.dev/layouts/stack/

Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Co-Authored-By: Itamar Givon <itamar.givon.dev@gmail.com>