Allow passing state from createAndMount to store
Instead of directly exporting a store instance, export a function to
create the store, which can pass through some initial state. In this
commit, we’re not actually defining any such state yet, but that will
come soon, so we already use the interface / object-destructure syntax
and temporarily silence the lints complaining about that.

Co-Authored-By: Noa wmde <noa.rave@wikimedia.de>
Co-Authored-By: Itamar Givon <itamar.givon@wikimedia.de>