Add lemma language input and submit form
This is a temporary solution to make creating lexemes work, by targeting
the noscript backend of the special page; we want to replace this with
an API request later. (At some point the hard-coded lemma-language will
also be replaced, of course.)

The previous simple Cypress test is replaced with one that watches for a
form submit and checks the submitted data (including a token that we
hard-code in index.html). Unfortunately, Chromium reports inexplicable
cross-origin errors when submitting, blocking the iframe with origin
http://localhost:3000 from accessing a cross-origin frame at
http://localhost:3000, or something like that (yes, that’s the same host
twice, as far as I can tell). I haven’t been able to find a solution for
that, so for now, work around it by disabling Chrome web security in
cypress.json. We can remove that line once we no longer submit the form.

Bug: T298153