Remove vue-tsc from build command
This command doesn’t actually contribute to the build results, it just
takes a lot of time (much longer than vite build itself). We already run
this as part of the test command, so I don’t think there’s a reason to
keep it in the build command as well.