Turn second createAndMount() param into services object
We’ll need more than one service soon, so let’s turn this into a
container of possibly several services. The messagesRepository service
remains optional for now, but the whole services object is required,
even if it’s just the empty object. (I’m not yet sure if additional
services will also be optional or required.)

Bug: T302961