Use the API to create lexemes
This wires up the new LexemeCreator service, used in a new action; to
avoid putting the service inside the store state, we turn the actions
into a function returning the actions object (instead of directly
exporting the object), so that we can pass the service into that
function and thus access it from the actions.

Some things are left for later: the lemma language code is hard-coded to
English, we don’t redirect to the lexeme’s page yet, and we don’t yet
pass any tags to the MwApiLexemeCreator either.

Bug: T302961
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>
Co-Authored-By: Noa Rave <noa.rave@wikimedia.de>