Add default parameters to mw.Api()
We want to make sure that the API uses the current user interface
language, even if that was only set via ?uselang= and isn’t in the
user’s session, and also want to use the modern result format. Also,
let’s set the error format already, though for now this is mostly a
guess since we haven’t implemented error handling yet.