Display error messages
Add an ErrorMessage component and use it in NewLexemeForm, showing the
first error from the store. The LexemeCreator can now reject with an
array of SubmitErrors, and the submit action saves these errors in the
store. The DevLexemeCreator can be made to produce an error by using the
special lemma “error”.

Bug: T304490
Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>