Add linting for BEM-style classes in scss
This effectively extends the functionality of the selector-class-pattern
rule, and so that rule can be dropped from the configuration.

From the other rule-sets mentioned in T301694,
stylelint-selector-bem-pattern has much more complex configuration for
little apparent benefit. And Vue-BEM adds a new custom directive that
also seems like too much complexity for what little of a problem we here
have.

Bug: T301694