PoC: Provide static HTML for info message
This extracts the info message into a component – note that we can’t use
the Wikit Message component as it is, because that puts all of its
content into a <span> (so we can only use phrasing content inside), so
instead we repeat some of the relevant styles – and uses that in the
app. Additionally, the component is rendered into a static HTML file
using a new build configuration (which emits just that component as a
separate common.js bundle) and a small script (render-info.js) using
Vue’s SSR renderer to generate the HTML.

This static HTML can then be copied into WikibaseLexeme.git and used by
the special page for the no-JS version. All the UPPER_SNAKE_CASE
placeholders in the component are intended to be replaced with messages
or other HTML fragments by the special page. Currently, the App
component will show those unmodified; I’m not yet sure if we want the
App component to render the info message component again (with proper
messages this time), or if it should just grab the replaced HTML that
the special page already has, so that the component would never be
rendered client-side.

Bug: T305445