PoC: Let caller inject static HTML for info message
This lets the caller include infoHtml in the config object, in which
case we’ll use that instead of mounting the InfoMessage component.

Also, make the InfoMessage fill its border box, since (due to the <div
v-html>) it’s no longer the direct descendant of the flex div.

The provide/inject between main and App should use some symbol as the
key, but I’m not sure where that symbol should live (<script setup> SFCs
*can* export things, but it’s a bit awkward).

Bug: T305445