Rename ListLanguageCodesProvider to MapLanguageCodesProvider
Since it now wraps a map, not a list (array).

While we’re at it, make the describe() in the test file name the class
being tested, rather than the interface. (But I think it makes sense to
still call the file LangCodeProvider.test.ts, since the file being
tested is also LangCodeProvider.ts, without List- or Map-.)