Don’t escape dev messages, only arguments
The messages themselves are hard-coded, we know they’re safe; this way,
we can also make the links in the shortcopyrightwarning message work.

This means that the NewLexemeForm integration tests now need to supply a
full config, otherwise `undefined` ends up being passed into the
messages plugin, where it now causes an error.