Mostly downgrade required Node version from 16 to 14
As far as I can tell, the history of the Node requirement so far was:

1. bc365a6229 (#10): CI is added, running on Node 16 (then-current LTS);
   .nvmrc=16 is added at the same time, presumably for consistency
2. c2aa5b7773 (#16): Node 16 added to README, presumably based on .nvmrc
3. e0c53b713d (#55): Node 16 added to package.json, based on .nvmrc

So the Node version requirement gradually, and I think unintentionally,
mutated from “let’s run CI on the latest software” into “we *need* the
latest software”. But a full `npm cit` run still works under Node 14,
and I think there is some value to supporting this version as well, as
it’s currently the latest Node version supported by Fresh [1]. So let’s
change most of the Node 16 requirements to Node 14 instead.

However, we do need npm v7 or later: npm v6 does not install peer
dependencies (even when they’re in the lockfile generated by a later npm
version!), which means linting fails because some PostCSS packages (peer
dependencies of the config we use) aren’t present. This isn’t a problem
in fresh-node14, which comes with npm v7; however, upstream Node 14
ships npm v6, so our .nvmrc still needs to contain 16, not 14, and in CI
(where we want to ensure compatibility with Node 14), we need to install
a newer npm version explicitly.

[1]: https://github.com/wikimedia/fresh

Bug: T302321