Change module type in tsconfig to commonjs
This should get rid of the following error:

Error: src/App.vue(1,1): error TS1323: Dynamic imports are only supported when the '--module' flag is set to 'es2020', 'es2022', 'esnext', 'commonjs', 'amd', 'system', 'umd', 'node12', or 'nodenext'.

I think this error is a false positive/bug as we do not actually have
any dynamic imports anywhere.

As I understand it, we do not use the output of the typescript
transpilation directly, but it is an intermediate product in the vite
build process.