Refactor ItemLookup to mostly use SearchedItemOption
Make most of ItemLookup deal with SearchedItemOption values instead of
WikitItemOptions (formerly MonolingualOption) values, and translate
between the two only in a section at the end of the script setup, at the
boundary to the Wikit component. When we migrate to Codex, this section
will hopefully become unnecessary, since Codex should properly support
multilingual options.