Don’t debounce item suggestions in ItemLookup
Move the handling of suggested items out of debouncedSearchForItems, and
instead make suggestedOptions computed based on itemSuggestions and
searchInput. (That requires moving searchInput up, which I’m not fully
happy with, to be honest.) Deduplicating the menu items, to avoid having
suggested items show up again in the search results, then happens in the
computed menuItems.