Don’t debounce item suggestions in ItemLookup
Move the handling of suggested items out of debouncedSearchForItems, and
instead make suggestedOptions computed based on itemSuggestions and
searchInput. (That requires moving searchInput up, which I’m not fully
happy with, to be honest.) Deduplicating the menu items, to avoid having
suggested items show up again in the search results, then happens in the
computed menuItems.

This fixes a minor bug in the previous commit: in onScroll(), the offset
is now finally again the correct length of previous search results,
neither including suggested items, nor excluding results returned by the
search but skipped as duplicates of suggested items.