Turn store language and lexicalCategory into SearchedItemOption
Change the value prop of the ItemLookup, LanguageInput and
LexicalCategoryInput components from a string (item ID) to a full
SearchedItemOption object, and also change the type in the store, so
that we use these objects (with an id, but also a display section)
throughout. This will let us populate the lookups from the URL
parameters.

Note that ItemLookup doesn’t fully react to incoming changes to the
value prop yet, that will come later.

Bug: T308118