Make ItemLookup more useful when given non-null value
If the value is changed from outside (e.g. when we load the language or
lexical category from URL parameters), it should be shown in the search
input as well, and when users click the input, there should be at least
one menu item rather than the “no results” message.

Bug: T308118