Change submit button while creating Lexeme
Change the button’s label, and disable it, while the Lexeme is being
created. When it’s done, reenable the button and change the label back.

Note that the button may still appear to be enabled when it’s clicked
with the mouse (instead of another way to submit the form, like pressing
Enter in the lemma input); this requires wmde/wikit#584 to fix.

In the tests, after rejecting the lexemeCreator promise, we apparently
need to wait for the next Vue tick six times. I don’t know why, but five
isn’t enough. I’m sorry.

Bug: T304511