Only reset submit button after navigation finishes
On the wiki, we don’t want the submit button to become enabled again
after the lexeme was created, while the user is being redirected there:
depending on network conditions and how fast the server responds, this
might still leave the button enabled for a perceptible duration.
Instead, on the wiki the button should really never be enabled again.
However, we want the dev entry point to be “reusable” without reloading
the page, so there the button should be enabled at some point.

Resolve this conflict by making the WikiRouter return a promise which
only resolves in the dev entry point, and chaining the reenabling of the
button after that. On the dev entry point, you can submit the form as
many times as you like; on the wiki, the promise never resolves, the
button stays disabled until the page is gone, and the user isn’t given
the opportunity to create the lexeme a second time.

Bug: T304511