Upgrade dependencies (#206)
There seems to be an issue with @vue/test-utils. As soon as I update that, I get type errors during testing. This needs some more investigation and then probably an issue opened there. Also, I think I would like to hold of on upgrading to jest 28 until that has been investigated a bit more.

Also, there seems to be a peer dependency conflict between `@vue/eslint-config-typescript` version 10 and `eslint-plugin-vue` version 9. This created an error in CI, which runs with node 14 and npm 8, though it did work locally with node 16 and npm 8. So this commit still includes `eslint-plugin-vue` version 8.

Bug: T309582