Don’t reset spelling variant when language stays the same
When the user tabs through the form, leaving an item lookup will emit a
“value” update for that lookup, even if the value didn’t change. That’s
fine for the lexical category lookup, but problematic for the language
lookup, because it would wipe whatever spelling variant the user had
already entered, and also retry looking up the language code of the
item. Instead, only have the HANDLE_LANGUAGE_CHANGE action reset the
relevant state if the language actually changed, not if it stayed the
same. (Note that we still commit the SET_LANGUAGE mutation, as the
display section of the item may have changed.)

Bug: T201587