Make ItemLookup searchInput a prop
This changes the ItemLookup’s searchInput from a ref (internal state) to
a prop (external state), and wires it up all the way to the store. This
is necessary so that the store can decide to show different error
messages for missing language / lexical category / spelling variant,
depending on whether the search input is empty (“missing”) or not
(“invalid”).

The responsibility for setting the search input to the item’s label (or
item ID) when the item is loaded via the URL moves from the ItemLookup
(watchEffect()) to the store (HANDLE_INIT_PARAMS action).

In the ItemLookup tests, we now need to set the input value and
searchInput prop at the same time. When the ItemLookup is properly used
in another component, it’s the outer component’s responsibility to
listen for update:searchInput events and change the prop accordingly; in
Vue Test Utils, it doesn’t seem to be possible to listen for specific
events, so we just set both directly in the setSearchInput() helper.

Bug: T310134