Make SpellingVariantInput searchInput a prop
Similar to what #223 does for the ItemLookup searchInput, make the
SpellingVariantInput searchInput a prop and wire it up with the store.
This will let the store show a different error message when the input
isn’t empty but hasn’t been selected (not implemented in this commit).

Also, make HANDLE_INIT_PARAMS set the search input when the spelling
variant has been loaded from the URL. Previously, the search input was
left empty in that case, making it appear as if the spelling variant
hadn’t been loaded. With this change, the input is set to just the
language code (rather than the language name and code, combined via an
interface message) – this should be improved further, but I think that
needs a bit more consideration.

Bug: T310134