Show warning for anonymous users
Using the standard Wikibase message, but without passing an argument
into it. Special:NewLexeme passes the wikibase-entity-lexeme message
into it (the argument used to be used as a sort of “your IP address will
be visible in the history of the item/property” thing), but that message
doesn’t actually exist; the $1 argument is also only used in one
remaining translation (vec/Venetian), where I’ve asked the translator to
update the translation so it doesn’t use $1 anymore. Once that’s done, I
expect we’ll update Wikibase and WikibaseLexeme to also drop this
message parameter.

On the dev entry point, the warning can be seen by adding
?isAnonymous=true to the URL.

Bug: T313466