Add aria-required to lookup inputs
We were already adding it to the lemma input; since Wikit 3.0.0-alpha.7,
we can also add it to the other three inputs, which are lookups.

In the ItemLookup component, the aria-required is passed through as a
new prop instead of hard-coded to true – all current item lookups are
required, but in principle the component could be used for a
non-required item lookup in future.

Note that we use :aria-required="true" instead of aria-required="true"
to pass `true` as a boolean instead of a string.

Bug: T313891