Style adjustments
The linter rule does not allow using 'rem' as a unit,
i've disabled it.
font-size for the copyright message should be 13px, the
rem in this case is relativly calculated from 1rem which is 16px.
margins between elements adapted to represent 16px as well.
The main content wrapper will get a top margin of 0.5rem
so that the gap between search-exiting or anonymous edit message
and the main form always remains 1.5rem.

Bug: T312652