Do not synthesize italic font in ToS&License copy (#331)
If the characters and font do not include italic variants, broken glyphs get rendered, which are particularly difficult to read in cursive scripts like those with Perso-Arabic alphabets. It is especially not a good idea here because the purpose of the NewLexeme page is explicitly multi-lingual.

Adjusting `font-synthesis` means that this paragraph is still in italics for fonts that have dedicated glyphs for that, but it is "roman" for fonts that do not.