Improve the required marker for fields
This moves the required marker closer to the actual label to make the
connection more obvious. Also it makes it bigger.

The style blocks for overriding the gaps must not be scoped, because it
otherwise the styles do not match anything.

The `line-height: 0;` has been added because otherwise the size of the
asterisk affects the whole label and thus also the overall layout.

Bug: T322683