Compare language codes case-insensitive
When switching to getting the language codes for a language from the
IETF-code property in T348923, we also get a lot of composite codes for
language variants that are partially uppercased, like en-GB or pt-BR.
However, in our data, they are stored as completely lowercase strings.
Thus, without this change, the user still has to manually, yet
needlessly, select the valid language code from the dropdown.