Lowercase language code from item statement
We now use the property “IETF language tag” for the automatic language
code on Wikidata; the subtag of these tags is case-insensitive, but some
capitalization is recommended [1][2]. For MediaWiki or Wikidata, on the
other hand, the subtags generally need to be all-lowercase (except our
special mis-x-Qid codes, see T317863). Reconcile this by lowercasing all
language codes from items in the store action.

[1]: https://en.wikipedia.org/wiki/IETF_language_tag#Syntax_of_language_tags
[2]: https://en.wikipedia.org/w/index.php?title=IETF_language_tag&oldid=1181524280#Syntax_of_language_tags

Bug: T349652