Process CentralAuth redirects on new lexeme as TempUser
With T357024, the API will return a URL initialized by CentralAuth
when a TempUser tries to create a new lexeme. We need to redirect
to this URL instead of the new lexeme (the URL will then redirect
to the new lexeme once it’s finished setting up the temporary
account).

Bug: T357152