Process CentralAuth redirects on new lexeme as TempUser (#640)
* Refactor WikiRouter to UrlLauncher

In preparation for changes required by T357024, move the
responsibility for generating the Wiki URL out of WikiRouter and
into the LexemeCreator. This way, the LexemeCreator always returns
an absolute URL (which in future may also be a redirect to another
site).

Bug: T357152

* Process CentralAuth redirects on new lexeme as TempUser

With T357024, the API will return a URL initialized by CentralAuth
when a TempUser tries to create a new lexeme. We need to redirect
to this URL instead of the new lexeme (the URL will then redirect
to the new lexeme once it’s finished setting up the temporary
account).

Bug: T357152