Use Codex Field + TextInput instead of Wikit TextInput
Change the lemma to not use Wikit anymore. In Codex, this corresponds to
two components, so use both.

The spacing of the “required” asterisk is done differently – it’s no
longer in a flex container, so establish the spacing using margin
instead. The asterisk is now also part of the actual <label> element,
which isn’t ideal but I think can’t be avoided within Codex. (Note that
Codex doesn’t want us to indicate required fields at all; see also
T374432.)

Bug: T370054