Move properties before nested selector to resolve deprecation warnings
SASS has been emitting a deprecation warning because of a coming
change in SASS behaviour:

https://sass-lang.com/documentation/breaking-changes/mixed-decls/

In `NewLexemeForm.vue` we were using the `&` parent selector
before applying additional properties to the `.wbl-snl-form` element.

In this specific case, moving the parent selector after the
property declarations has no effect on the semantics of the SASS or
the generated CSS, and resolves the deprecation warning.

Bug: T372250